+++ build.gradle
@@ -28,38 +28,38 @@ def doOnChange = './gradlew classes'
 logging.captureStandardOutput LogLevel.INFO
 
 dependencies {
-  compile "io.vertx:vertx-core:$vertxVersion"
-  compile "io.vertx:vertx-config:$vertxVersion"
-  compile "io.vertx:vertx-web:$vertxVersion"
-  compile "io.vertx:vertx-web-api-contract:$vertxVersion"
-  compile "io.vertx:vertx-web-client:$vertxVersion"
-  compile "io.github.classgraph:classgraph:4.4.12"
+  implementation "io.vertx:vertx-core:$vertxVersion"
+  implementation "io.vertx:vertx-config:$vertxVersion"
+  implementation "io.vertx:vertx-web:$vertxVersion"
+  implementation "io.vertx:vertx-web-api-contract:$vertxVersion"
+  implementation "io.vertx:vertx-web-client:$vertxVersion"
+  implementation "io.github.classgraph:classgraph:4.4.12"
 
   implementation 'org.apache.httpcomponents.client5:httpclient5:5.0.1'
   implementation 'org.apache.httpcomponents.client5:httpclient5-fluent:5.0.1'
 
 
-  compile name: "cartago", version: "3.1-SNAPSHOT"
-  compile name: "wot-td-java", version: "0.0.1"
-  compile group: "org.jacamo", name: "moise", version: "1.0"
+  implementation name: "cartago", version: "3.1-SNAPSHOT"
+  implementation name: "wot-td-java", version: "0.0.1"
+  implementation group: "org.jacamo", name: "moise", version: "1.0"
 
 
   // https://mvnrepository.com/artifact/org.eclipse.rdf4j/rdf4j-model
-  compile group: 'org.eclipse.rdf4j', name: 'rdf4j-model', version: '2.4.2'
+  implementation group: 'org.eclipse.rdf4j', name: 'rdf4j-model', version: '2.4.2'
   // https://mvnrepository.com/artifact/org.eclipse.rdf4j/rdf4j-repository-sail
-  compile group: 'org.eclipse.rdf4j', name: 'rdf4j-repository-sail', version: '2.4.2'
+  implementation group: 'org.eclipse.rdf4j', name: 'rdf4j-repository-sail', version: '2.4.2'
   // https://mvnrepository.com/artifact/org.eclipse.rdf4j/rdf4j-sail-memory
-  compile group: 'org.eclipse.rdf4j', name: 'rdf4j-sail-memory', version: '2.4.2'
+  implementation group: 'org.eclipse.rdf4j', name: 'rdf4j-sail-memory', version: '2.4.2'
 
-  compile group: 'com.mbed.java-coap', name: 'coap-core', version: '5.0.0'
+  implementation group: 'com.mbed.java-coap', name: 'coap-core', version: '5.0.0'
 
-  compile 'org.apache.commons:commons-rdf-api:0.5.0'
-  compile 'org.apache.commons:commons-rdf-rdf4j:0.5.0'
-  compile 'com.google.code.gson:gson:2.8.6'
+  implementation 'org.apache.commons:commons-rdf-api:0.5.0'
+  implementation 'org.apache.commons:commons-rdf-rdf4j:0.5.0'
+  implementation 'com.google.code.gson:gson:2.8.6'
 
   implementation 'ch.qos.logback:logback-classic:1.2.9'
-  testCompile "junit:junit:4.12"
-  testCompile "io.vertx:vertx-unit:$vertxVersion"
+  testImplementation "junit:junit:4.12"
+  testImplementation "io.vertx:vertx-unit:$vertxVersion"
   //testImplementation 'ch.qos.loback:logback-classic:1.2.9'
 }
