+++ initialization_uc1_steyr.sh
@@ -1,0 +1,1 @@
+./initialization_uc1.sh --hyper http://localhost:8080 --avl  http://192.168.55.1 --hil http://192.168.88.28:8545
