+++ initialization_uc1_vienna.sh
@@ -1,1 +1,1 @@
-./initialization.sh --hyper http://192.168.55.100:8081 --avl  192.168.55.1 --hil http://192.168.55.1:8089/intelliot/uc1-hil-service
+./initialization_uc1.sh --hyper http://192.168.55.100:8081 --avl  192.168.55.1 --hil http://192.168.55.1:8089/intelliot/uc1-hil-service
