--- src/main/java/org/hyperagents/yggdrasil/cartago/EventManagerCallback.java
@@ -1,20 +1,0 @@
-package org.hyperagents.yggdrasil.cartago;
-
-import cartago.*;
-
-
-public class EventManagerCallback implements ICartagoCallback {
-
-  EventManager manager;
-
-
-  public EventManagerCallback(EventManager manager){
-    this.manager = manager;
-  }
-
-  @Override
-  public void notifyCartagoEvent(CartagoEvent ev) {
-    manager.add(ev);
-
-  }
-}
