+++ src/main/java/org/hyperagents/yggdrasil/jason/RunYggdrasilMAS.java
@@ -2,9 +2,7 @@ package org.hyperagents.yggdrasil.jason;
 
 import jason.architecture.AgArch;
 import jason.architecture.MindInspectorAgArch;
-import jason.infra.centralised.CentralisedAgArch;
 import jason.infra.centralised.RunCentralisedMAS;
-import jason.runtime.RuntimeServicesFactory;
 
 public class RunYggdrasilMAS extends RunCentralisedMAS {
 
@@ -21,14 +19,5 @@ public class RunYggdrasilMAS extends RunCentralisedMAS {
     AgArch agArch = getDFAgArch();
     System.out.println("df ag: "+ agArch.getAgName());
   }
-
-  /*@Override
-  public void addAg(CentralisedAgArch agArch) throws IllegalArgumentException{
-    if (agArch instanceof YggdrasilAgArch){
-      super.addAg(agArch);
-    } else {
-      throw new IllegalArgumentException();
-    }
-  }*/
 
 }
