--- src/main/java/org/hyperagents/yggdrasil/jason/YggdrasilAgArch.java
@@ -1,17 +1,0 @@
-package org.hyperagents.yggdrasil.jason;
-
-//import jaca.CAgentArch;
-import jason.infra.centralised.CentralisedAgArch;
-
-public class YggdrasilAgArch extends CentralisedAgArch {
-
-
-
-  @Override
-  public void init(){
-
-    //this.insertAgArch(new CAgentArch());
-    //System.out.println("insert YAgentArch");
-    //this.insertAgArch(new YAgentArch());
-  }
-}
