+++ start_localhost.sh
@@ -1,0 +1,1 @@
+java -jar build/libs/yggdrasil-0.0-SNAPSHOT-fat.jar -conf src/main/conf/config.json
