+++ uc1_hs_steyr.sh
@@ -1,0 +1,1 @@
+./uc1_hs.sh --agent manager --hyper http://localhost:8080
